add verification

+++ controleur/commentaire.js
@@ -1,8 +1,12 @@
-const commentaire = require('../modele/commentaire')
+const commentaire = require('../modele/commentaire');
+const utilisateur = require('../modele/user');
+const post = require('../modele/post')
 
 exports.creerComentaire = (req, res, next) => {
     commentaire.create({
-        contenu: req.body.contenu
+        contenu: req.body.contenu,
+        user_id: req.body.user_id,
+        post_id: req.body.post_id
     })
     .then(() => res.status(201).json({ message: 'commentaire créer' }))
     .catch(erreur => res.status(400).json({ erreur }))
@@ -25,23 +29,28 @@ exports.modifierUnCommentaire = (req, res, next) => {
 exports.supprimerUnCommentaire = (req, res, next) => {
     commentaire.findOne({ where: {id: req.params.id} })
     .then(() => {
-        commentaire.destroy({ where: {id: req.params.id} })
-        .then(() => {
-            res.status(201).json({ message: 'commentaire supprimé' })
-        })
-        .catch(error => res.status(404).json({ error }));
+        if(req.body.admin == true) {
+            commentaire.destroy({ where: {id: req.params.id} })
+            .then(() => {
+                res.status(201).json({ message: 'commentaire supprimé' })
+            })
+            .catch(error => res.status(404).json({ error }));
+        } else {
+            return res.status(400).json({ message: "Seul un administrateur peut supprimer les commentaires ! "})
+        }
+        
     })
     .catch(error => res.status(404).json({ error }));
 }
 
 exports.afficherUnCommentaire = (req, res, next) => {
-    commentaire.findOne({ where: {id: req.params.id} })
-    .then(commentaire => res.status(201).json({ commentaire }))
+    commentaire.findAll({include: [{model: post, where: {id: req.params.id}}, {model: utilisateur} ]})
+    .then(commentaire => res.status(200).json({ commentaire }))
     .catch(error => res.status(404).json({ error }));
 }
 
 exports.afficherTousLesCommentaire = (req, res, next) => {
     commentaire.findAll()
-    .then(commentaire => res.status(201).json({ commentaire }))
+    .then(commentaire => res.status(200).json({ commentaire }))
     .catch(error => res.status(404).json({ error }));
 }
